+++ module/herosystem6e.js
@@ -649,6 +649,7 @@ Hooks.on('updateWorldTime', async (worldTime, options, userId) => {
 
         // Out of combat recovery.  When SimpleCalendar is used to advance time.
         // This simple routine only handles increments of 12 seconds or more.
+        const automation = game.settings.get("hero6efoundryvttv2", "automation");
         if ((automation === "all") || (automation === "npcOnly" && actor.type == 'npc') || (automation === "pcEndOnly" && actor.type === 'pc')) {
             if ((parseInt(options) || 0) >= 12 && (
                 parseInt(actor.system.characteristics.end.value) < parseInt(actor.system.characteristics.end.max) ||
@@ -660,6 +661,32 @@ Hooks.on('updateWorldTime', async (worldTime, options, userId) => {
                 actor.system.characteristics.end.value = Math.min(parseInt(actor.system.characteristics.end.max), parseInt(actor.system.characteristics.end.value) + rec)
                 actor.system.characteristics.stun.value = Math.min(parseInt(actor.system.characteristics.stun.max), parseInt(actor.system.characteristics.stun.value) + rec)
                 actor.update({ 'system.characteristics.end.value': actor.system.characteristics.end.value, 'system.characteristics.stun.value': actor.system.characteristics.stun.value }, { 'render': true })
+            }
+        }
+
+        // Charges Recover each day
+        if ((parseInt(options) || 0) >= 86400) {
+            const itemsWithCharges = actor.items.filter(o => o.system.charges?.max);
+            let content = "";
+            for (let item of itemsWithCharges) {
+                let value = parseInt(item.system.charges.value);
+                let max = parseInt(item.system.charges.max);
+                if (value < max) {
+                    content += `${actor.name}/${item.name} ${value} to ${max} charges.  `;
+                    item.update({ 'system.charges.value': max })
+                }
+
+            }
+
+            if (content) {
+                const chatData = {
+                    user: game.user.id, //ChatMessage.getWhisperRecipients('GM'),
+                    whisper: ChatMessage.getWhisperRecipients("GM"),
+                    speaker: ChatMessage.getSpeaker({ actor: this }),
+                    blind: true,
+                    content: content,
+                }
+                await ChatMessage.create(chatData)
             }
         }
     }